# Improve DirectoryTree & FileFinder_RTP performance