As observed by @BlisterB and @carstene1ns the RTP detection is incredible slow on Android but even on other systems we are wasting time in the detector.

Added three layers of _speed_ to make the RTP detection on startup faster:

1. When the RTP folder is empty the folder is considered invalid and not scanned (Android will hit this one)
2. When not empty the RTP-type-scan will abort when for a certain RTP type there are 10 missing files (this makes RTP2k detect ``30%`` faster, RTP2k3 detect ``59%`` faster).
3. When a folder is missing this is cached making the lookup much faster  because the filesystem is not hit again (this makes the ``RTP::Detect`` Unit Test where almost all folders are missing and "abort after 10 misses" is disabled ``20%`` faster). On slow systems with shitty IPC (hello Android and 3DS) the speedup in this rare error case should be gigantic.

---------

Redesigned the data structure of the DirectoryTree from std::unordered_map to sorted std::vector as suggested by fmatthew in #2306

This gives a ``5%`` speedup for ``FindFile`` lookups in my benchmarks and reduces memory usage of the tree by ``15%``. Could be more on systems with a smaller cpu cache.

---------

Faster startup when a RTP exists (or the folder is empty) should be noticable on slower systems. Faster ``FindFile`` likely not (microbenchmarks...).

---------

Android loading times (old):

- RTP Off: 2s 
- RTP On & installed: 5s
- RTP On & empty: 16s

Android loading times this branch: _Note that the PR build has debug on so you will not hit these times_

- RTP Off: 2s
- RTP On & installed: 5s (could be a few ms faster but can't properly measure this on Android)
- RTP On & empty: **2s**