@Ghabry 2022-06-10
@BlisterB I think the times are good enough to have RTP enabled by default now. It is still few seconds slower (when installed) but when not installed it will not impact startup.