@Ghabry 2022-06-10
I could delay the RTP check until the first asset is requested.
Then you would get a hang while in-game but only for RTP games. Would this be preferred?