@carstene1ns 2022-06-10
I think doing the RTP detection while the game is running will involve stutters in audio. So, merging.